Fix S3M effects - period should clamp to 64 at minimum - add various test songs
